services/ds160-processor/requirements.txt
pandas==1.5.3
playwright==1.42.0
requests==2.31.0
pdf2image==1.16.3
Pillow==10.1.0 